requirements.txt
accelerate==0.26.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
betterproto==2.0.0b6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cramjam==2.9.0
cryptography==44.0.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
distributed==2025.3.0
dnspython==2.7.0
einops==0.8.0
email_validator==2.2.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.11.0
fasttext==0.9.3
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
grpclib==0.4.7
h2==4.2.0
hpack==4.1.0
huggingface-hub==0.26.2
hyperframe==6.1.0
id==1.5.0
idna==3.10
importlib-resources==5.13.0
importlib_metadata==8.6.1
in-toto-attestation==0.9.3
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kagglehub==0.3.7
kiwisolver==1.4.7
llvmlite==0.44.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
model-signing==0.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rfc3161-client==0.1.2
rfc8785==0.1.4
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
securesystemslib==1.2.0
sentence-transformers==3.3.1
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.16.0
smart-open==7.0.5
sortedcontainers==2.4.0
sparse==0.16.0
stack-data==0.6.3
sympy==1.13.1
tblib==3.1.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
tuf==5.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==2.1.2
xxhash==3.5.0
yarl==1.18.0
zict==3.0.0
zipp==3.21.0
